pyproject.toml
[tool.poetry]
name = "infobot"
version = "0.1.0"
description = "智能微信機器人 - 班長助手，用於自動化處理老師信息並轉發給學生"
authors = ["Icarus <zhehongl91@gmail.com>"]
readme = "README.md"
packages = [{include = "infobot"}]

[tool.poetry.dependencies]
python = "^3.10.17"
requests = "^2.31.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
loguru = "^0.7.2"
schedule = "^1.2.0"
pyyaml = "^6.0.1"
pyobjc-framework-cocoa = "^10.1"
pyobjc-framework-quartz = "^10.1"
pyobjc-framework-applicationservices = "^10.1"
pillow = "^10.1.0"
opencv-python = "^4.8.1.78"
numpy = "^1.26.0"
pyautogui = "^0.9.54"
pygetwindow = "^0.0.9"
pywinctl = "^0.4"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88 